# BankingSystemAPIs/apidocs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Banking System API Documentation",
        "description": "The Banking System API provides a comprehensive set of endpoints for managing banking-related operations. It is designed to support various banking activities such as account management, transaction processing, and exchange rate management. This API is built with scalability and security in mind, ensuring seamless integration with client applications.\r\n\r\nKey Features:\r\nAccount Management: Create, retrieve, update, and delete accounts. Perform operations such as deposits, withdrawals, and transfers.\r\nTransaction Processing: Retrieve transaction histories, manage transaction records, and delete transactions securely.\r\nExchange Rates: Manage exchange rates for supported currencies, including retrieval, creation, updates, and deletions.\r\nLogging: Detailed request and response logging to facilitate debugging and auditing.\r\nRESTful Design: Follows RESTful principles with predictable endpoints, standard HTTP methods, and meaningful status codes.\r\nSwagger Documentation: Integrated Swagger/OpenAPI support for testing and visualizing API capabilities.",
        "contact": {
            "name": "Mahmoud Metwalli",
            "email": "mammkbih@gmail.com"
        },
        "version": "v1"
    },
    "paths": {
...
        "/api/accounts/{id}/balance": {
            "get": {
                "tags": [
                    "Main"
                ],
                "summary": "Retrieves the balance of a specific account.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique identifier of the account.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/accounts/deposit": {
            "post": {
                "tags": [
                    "Main"
                ],
                "summary": "Deposits a specified amount into an account.",
                "requestBody": {
                    "description": "The deposit request details.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepositRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepositRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepositRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/accounts/withdraw": {
            "post": {
                "tags": [
                    "Main"
                ],
                "summary": "Withdraws a specified amount from an account.",
                "requestBody": {
                    "description": "The withdrawal request details.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WithdrawRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WithdrawRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/WithdrawRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/accounts/transfer": {
            "post": {
                "tags": [
                    "Main"
                ],
                "summary": "Transfers a specified amount from one account to another.",
                "requestBody": {
                    "description": "The transfer request details.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransferRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransferRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransferRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DepositRequest": {
                "required": [
                    "accountId",
                    "amount",
                    "currency"
                ],
                "type": "object",
                "properties": {
                    "accountId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "amount": {
                        "minimum": 0.01,
                        "type": "number",
                        "format": "double"
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": {}
            },
...
            "TransferRequest": {
                "required": [
                    "accountId",
                    "amount",
                    "currency",
                    "targetAccountId"
                ],
                "type": "object",
                "properties": {
                    "accountId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "targetAccountId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "amount": {
                        "minimum": 0.01,
                        "type": "number",
                        "format": "double"
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
            "WithdrawRequest": {
                "required": [
                    "accountId",
                    "amount",
                    "currency"
                ],
                "type": "object",
                "properties": {
                    "accountId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "amount": {
                        "minimum": 0.01,
                        "type": "number",
                        "format": "double"
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    }
                },
                "additionalProperties": false
            }
        }
    }
}